Move submission portal types/interfaces/consts to separate module
