+++ web/src/views/SubmissionPortal/store/api.ts
@@ -1,17 +1,15 @@
 import { AxiosResponse } from 'axios';
-import { client, SearchParams, User } from '@/data/api';
-import { HARMONIZER_TEMPLATES, MetadataSuggestion, MetadataSuggestionRequest } from '../harmonizerApi';
-
-interface NmdcAddress {
-  name: string;
-  email: string;
-  phone: string;
-  line1: string;
-  line2: string;
-  city: string;
-  state: string;
-  postalCode: string;
-}
+import { client, SearchParams } from '@/data/api';
+import {
+  LockOperationResult,
+  MetadataSubmission,
+  MetadataSubmissionRecord,
+  MetadataSuggestion,
+  MetadataSuggestionRequest,
+  NmdcAddress,
+  PaginatedResponse,
+  SuggestionType,
+} from '@/views/SubmissionPortal/types';
 
 function addressToString(address: NmdcAddress): string {
   let result = '';
@@ -25,42 +23,6 @@ function addressToString(address: NmdcAddress): string {
   const joinString = (address.city.trim() && stateAndZip.trim()) ? ', ' : '';
   result += [address.city, stateAndZip].join(joinString);
   return result;
-}
-
-interface MetadataSubmission {
-  packageName: (keyof typeof HARMONIZER_TEMPLATES)[];
-  contextForm: any;
-  addressForm: any;
-  templates: string[];
-  studyForm: any;
-  multiOmicsForm: any;
-  sampleData: Record<string, any[]>;
-}
-
-interface MetadataSubmissionRecord {
-  id: string;
-  author_orcid: string;
-  created: string;
-  metadata_submission: MetadataSubmission;
-  status: string;
-  locked_by: User;
-  lock_updated: string;
-  permission_level: string | null;
-  source_client: 'submission_portal' | 'field_notes' | 'nmdc_edge' | null;
-  study_name: string;
-  templates: string[];
-}
-
-interface PaginatedResponse<T> {
-  count: number;
-  results: T[];
-}
-
-interface LockOperationResult {
-  success: boolean;
-  message: string
-  locked_by?: User | null;
-  lock_updated?: string | null;
 }
 
 async function createRecord(record: MetadataSubmission) {
@@ -126,10 +88,7 @@ async function getMetadataSuggestions(data: MetadataSuggestionRequest[]) {
 }
 
 export {
-  NmdcAddress,
   addressToString,
-  MetadataSubmission,
-  MetadataSubmissionRecord,
   createRecord,
   getRecord,
   listRecords,
