+++ web/src/views/SubmissionPortal/store/index.ts
@@ -5,41 +5,36 @@ import CompositionApi, {
 import { clone, forEach } from 'lodash';
 import axios from 'axios';
 import * as api from './api';
-import { getVariants, HARMONIZER_TEMPLATES, MetadataSuggestion } from '../harmonizerApi';
+import { getVariants } from '../harmonizerApi';
 import { User } from '@/data/api';
+import {
+  HARMONIZER_TEMPLATES,
+  MetadataSubmission,
+  MetadataSuggestion,
+  NmdcAddress,
+  PermissionLevelValues,
+  PermissionTitle,
+  SubmissionStatus,
+  SuggestionType,
+  SuggestionsMode,
+} from '@/views/SubmissionPortal/types';
 
 // TODO: Remove in version 3;
 Vue.use(CompositionApi);
 
-enum BiosafetyLevels {
-  BSL1 = 'BSL1',
-  BSL2 = 'BSL2'
-}
-
-enum AwardTypes {
-  CSP = 'CSP',
-  BERSS = 'BERSS',
-  BRCS = 'BRCs',
-  MONET = 'MONet',
-  FICUS = 'FICUS'
-}
-
-type permissionTitle = 'Viewer' | 'Metadata Contributor' | 'Editor';
-type permissionLevelValues = 'viewer' | 'metadata_contributor' | 'editor' | 'owner';
-const permissionTitleToDbValueMap: Record<permissionTitle, permissionLevelValues> = {
+const permissionTitleToDbValueMap: Record<PermissionTitle, PermissionLevelValues> = {
   Viewer: 'viewer',
   'Metadata Contributor': 'metadata_contributor',
   Editor: 'editor',
 };
 
-const permissionLevelHierarchy: Record<permissionLevelValues, number> = {
+const permissionLevelHierarchy: Record<PermissionLevelValues, number> = {
   owner: 4,
   editor: 3,
   metadata_contributor: 2,
   viewer: 1,
 };
 
-type SubmissionStatus = 'In Progress' | 'Submitted- Pending Review' | 'Complete';
 const submissionStatus: Record<string, SubmissionStatus> = {
   InProgress: 'In Progress',
   SubmittedPendingReview: 'Submitted- Pending Review',
@@ -58,8 +53,8 @@ function getSubmissionLockedBy(): User | null {
   return _submissionLockedBy;
 }
 
-let _permissionLevel: permissionLevelValues | null = null;
-function getPermissionLevel(): permissionLevelValues | null {
+let _permissionLevel: PermissionLevelValues | null = null;
+function getPermissionLevel(): PermissionLevelValues | null {
   return _permissionLevel;
 }
 
@@ -93,7 +88,7 @@ const addressFormDefault = {
     city: '',
     state: '',
     postalCode: '',
-  } as api.NmdcAddress,
+  } as NmdcAddress,
   expectedShippingDate: undefined as undefined | Date,
   shippingConditions: '',
   // Sample info
@@ -138,7 +133,7 @@ const studyFormDefault = {
     name: string;
     orcid: string;
     roles: string[];
-    permissionLevel: permissionLevelValues | null;
+    permissionLevel: PermissionLevelValues | null;
   }[],
 };
 const studyFormValid = ref(false);
@@ -193,12 +188,6 @@ const templateChoiceDisabled = computed(() => {
   return false;
 });
 const metadataSuggestions = ref([] as MetadataSuggestion[]);
-
-enum SuggestionsMode {
-  LIVE = 'Live',
-  ON_DEMAND = 'On Demand',
-  OFF = 'Off',
-}
 const suggestionMode = ref(SuggestionsMode.LIVE);
 
 const tabsValidated = ref({} as Record<string, boolean>);
@@ -211,7 +200,7 @@ watch(templateList, () => {
 });
 
 /** Submit page */
-const payloadObject: Ref<api.MetadataSubmission> = computed(() => ({
+const payloadObject: Ref<MetadataSubmission> = computed(() => ({
   packageName: packageName.value,
   contextForm,
   addressForm,
@@ -221,8 +210,8 @@ const payloadObject: Ref<api.MetadataSubmission> = computed(() => ({
   sampleData: sampleData.value,
 }));
 
-function getPermissions(): Record<string, permissionLevelValues> {
-  const permissions: Record<string, permissionLevelValues> = {};
+function getPermissions(): Record<string, PermissionLevelValues> {
+  const permissions: Record<string, PermissionLevelValues> = {};
   studyForm.contributors.forEach((contributor) => {
     const { orcid, permissionLevel } = contributor;
     if (orcid && permissionLevel) {
@@ -271,8 +260,8 @@ async function incrementalSaveRecord(id: string): Promise<number | void> {
     return Promise.resolve();
   }
 
-  let payload: Partial<api.MetadataSubmission> = {};
-  let permissions: Record<string, permissionLevelValues> | undefined;
+  let payload: Partial<MetadataSubmission> = {};
+  let permissions: Record<string, PermissionLevelValues> | undefined;
   if (isOwner()) {
     payload = payloadObject.value;
     permissions = getPermissions();
@@ -311,7 +300,7 @@ async function loadRecord(id: string) {
   sampleData.value = val.metadata_submission.sampleData;
   hasChanged.value = 0;
   status.value = isSubmissionStatus(val.status) ? val.status : submissionStatus.InProgress;
-  _permissionLevel = (val.permission_level as permissionLevelValues);
+  _permissionLevel = (val.permission_level as PermissionLevelValues);
 
   try {
     const lockResponse = await api.lockSubmission(id);
@@ -355,14 +344,8 @@ function addMetadataSuggestions(suggestions: MetadataSuggestion[]) {
 }
 
 export {
-  SubmissionStatus,
   submissionStatus,
-  BiosafetyLevels,
-  AwardTypes,
-  SuggestionsMode,
-  permissionTitle,
   permissionTitleToDbValueMap,
-  permissionLevelValues,
   permissionLevelHierarchy,
   /* state */
   multiOmicsForm,
